requirements.txt
telebot
pillow
requests
aiohttp
python-dotenv
Pathlib